Add files via upload
Designed a text-based adventure game called Survival of the smartest. This game was developed using Python essentials utilizing String Manipulation, User-Defined Functions and Variables, Conditional and Nested Conditional Statements (if/elif/else) and Loops. 
The game is an adapted version of the paper-based game called 'Hangman'. The game is simple and requires you to guess the letters of the word and reveal it, to clear each level without dying. You will be given 8 chances to guess the word right. For every wrong guess a part of the body is drawn and upon losing all your chances will be lost. You will see a man hanged, at the end of it, or in other words, you lost your adventure.